+++ src/main/java/com/raoulvdberge/refinedstorage/apiimpl/autocrafting/craftingmonitor/CraftingMonitorElementFluidRender.java
@@ -29,7 +29,7 @@ public class CraftingMonitorElementFluidRender implements ICraftingMonitorElemen
     public void draw(int x, int y, IElementDrawers drawers) {
         drawers.getFluidDrawer().draw(x + 2 + offset, y + 1, stack);
 
-        float scale = 0.5f;
+        float scale = drawers.getFontRenderer().getUnicodeFlag() ? 1F : 0.5F;
 
         GlStateManager.pushMatrix();
         GlStateManager.scale(scale, scale, 1);