+++ src/main/java/com/raoulvdberge/refinedstorage/gui/GuiBase.java
@@ -6,6 +6,7 @@ import com.raoulvdberge.refinedstorage.api.render.IElementDrawers;
 import com.raoulvdberge.refinedstorage.gui.sidebutton.SideButton;
 import com.raoulvdberge.refinedstorage.inventory.ItemHandlerFluid;
 import com.raoulvdberge.refinedstorage.util.RenderUtils;
+import net.minecraft.client.gui.FontRenderer;
 import net.minecraft.client.gui.GuiButton;
 import net.minecraft.client.gui.inventory.GuiContainer;
 import net.minecraft.client.renderer.GlStateManager;
@@ -48,6 +49,11 @@ public abstract class GuiBase extends GuiContainer {
         @Override
         public IElementDrawer<String> getStringDrawer() {
             return GuiBase.this::drawString;
+        }
+
+        @Override
+        public FontRenderer getFontRenderer() {
+            return fontRenderer;
         }
     }